requirements.txt
asn1crypto==1.4.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10k
cryptography==36.0.1
google-api-core==2.5.0
google-auth==2.6.0
google-cloud==0.34.0
google-cloud-pubsub==2.9.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
libcst==0.4.1
mypy-extensions==0.4.3
numpy==1.22.1
oscrypto==1.2.1
pandas==1.4.0
proto-plus==1.20.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.13.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
snowflake-connector-python==2.7.3
typing-extensions==4.0.1
typing-inspect==0.7.1
urllib3==1.26.8
